refactor: Move ReflectUtils missdeclared method to ClassUtils
